refactor: enhance html semantic by refactoring some components and changing their tags
