+++ src/components/Footer/Footer.jsx
@@ -1,0 +1,17 @@
+import React from 'react'
+
+import { Footer as PlurallFooter } from 'plurall-footer'
+
+import styles from './Footer.module.scss'
+
+const Footer = () => {
+  return (
+    <footer className={styles.footer}>
+      <PlurallFooter />
+    </footer>
+  )
+}
+
+Footer.propTypes = {}
+
+export default Footer
