+++ src/components/Footer/Footer.module.scss
@@ -1,0 +1,4 @@
+.footer {
+  background-color: var(--color-dark-gray);
+  padding: var(--padding);
+}
