+++ src/components/Header/Header.module.scss
@@ -1,0 +1,3 @@
+.navBar {
+  background-color: var(--color-purple);
+}
