+++ src/components/Layout/Layout.module.scss
@@ -1,12 +1,3 @@
-.navBar {
-  background-color: var(--color-purple);
-}
-
 .content {
   min-height: calc(100vh - var(--height-nav-bar) - var(--height-footer));
-}
-
-.footer {
-  background-color: var(--color-dark-gray);
-  padding: var(--padding);
 }
