+++ src/components/SubHeader/SubHeader.jsx
@@ -5,22 +5,24 @@ import { ArrowButton, Breadcrumb, Heading, Link } from 'plurall-ui'
 
 import styles from './SubHeader.module.scss'
 
-const SubHeader = ({ buttonHref, breadcrumb, heading }) => (
-  <div className={styles.header}>
-    <div className={styles.wrapper}>
-      {buttonHref && (
-        <Link href={buttonHref} className={styles.button}>
-          <ArrowButton direction="left" />
-        </Link>
-      )}
+const SubHeader = ({ buttonHref, breadcrumb, heading }) => {
+  return (
+    <header className={styles.header}>
+      <div className={styles.wrapper}>
+        {buttonHref && (
+          <Link href={buttonHref} className={styles.button}>
+            <ArrowButton direction="left" />
+          </Link>
+        )}
 
-      <div className={styles.innerDiv}>
-        <Breadcrumb content={breadcrumb} className={styles.breadcrumb} />
-        <Heading className={styles.heading}>{heading}</Heading>
+        <div className={styles.innerDiv}>
+          <Breadcrumb content={breadcrumb} className={styles.breadcrumb} />
+          <Heading className={styles.heading}>{heading}</Heading>
+        </div>
       </div>
-    </div>
-  </div>
-)
+    </header>
+  )
+}
 
 SubHeader.propTypes = {
   breadcrumb: PropTypes.arrayOf(
