+++ src/components/index.js
@@ -1,4 +1,6 @@
 export { default as Layout } from './Layout'
+export { default as Header } from './Header'
+export { default as Footer } from './Footer'
 export { default as SubHeader } from './SubHeader'
 export { default as PrivateRoute } from './PrivateRoute'
 export { default as ArtistsList } from './ArtistsList'
